Fix the warm-reboot script to support FRR based warm-reboot
In case we use FRR stack with watchfrr, the watchfrr should be killed
during the warm-reboot shutdown process. This will make sure that watchfrr
won't restart the zebra/bgpd etc to screw up the warm-reboot process.

The code support the case where watchfrr is not used like quagga.

Signed-off-by: Zhenggen Xu <zxu@linkedin.com>

RB=2012276
G=lnos-reviewers
R=pchaudhary,pmao,vapatil,samaity,zxu
A=