[config/main.py]: Do not run destructor if lock was never acquired.
We have lock instance creation at global level, so that destructor can
run as soon as process is finished. And we acquire lock only when
config command is executed, not with TAB or config ?.

In this case, config ? has not acquired the lock but when process is
finished the destructor tried to release the lock. To solve this,
we can check if client, a class level variable was initialized or not.

Signed-off-by: Praveen Chaudhary pchaudhary@linkedin.com
RB=2249613
G=lnos-reviewers
R=pchaudhary,pmao,rmolina,samaity,sfardeen,zxu
A=